Put logic for new API methods in dedicated class
Calculate severity if NULL in database

Adjust tests
