+++ src/main/java/org/dependencytrack/persistence/QueryManager.java
@@ -110,6 +110,8 @@ public class QueryManager extends AlpineQueryManager {
     private CacheQueryManager cacheQueryManager;
     private ComponentQueryManager componentQueryManager;
     private FindingsQueryManager findingsQueryManager;
+
+    private FindingsSearchQueryManager findingsSearchQueryManager;
     private LicenseQueryManager licenseQueryManager;
     private MetricsQueryManager metricsQueryManager;
     private NotificationQueryManager notificationQueryManager;
@@ -275,6 +277,17 @@ public class QueryManager extends AlpineQueryManager {
             findingsQueryManager = (request == null) ? new FindingsQueryManager(getPersistenceManager()) : new FindingsQueryManager(getPersistenceManager(), request);
         }
         return findingsQueryManager;
+    }
+
+    /**
+     * Lazy instantiation of FindingsSearchQueryManager.
+     * @return a FindingsSearchQueryManager object
+     */
+    private FindingsSearchQueryManager getFindingsSearchQueryManager() {
+        if (findingsSearchQueryManager == null) {
+            findingsSearchQueryManager = (request == null) ? new FindingsSearchQueryManager(getPersistenceManager()) : new FindingsSearchQueryManager(getPersistenceManager(), request);
+        }
+        return findingsSearchQueryManager;
     }
 
     /**
@@ -1034,11 +1047,11 @@ public class QueryManager extends AlpineQueryManager {
     }
 
     public List<Finding> getAllFindings(final Map<String, String> filters, final boolean showSuppressed, final boolean showInactive) {
-        return getFindingsQueryManager().getAllFindings(filters, showSuppressed, showInactive);
+        return getFindingsSearchQueryManager().getAllFindings(filters, showSuppressed, showInactive);
     }
 
     public List<GroupedFinding> getAllFindingsGroupedByVulnerability(final Map<String, String> filters, final boolean showInactive) {
-        return getFindingsQueryManager().getAllFindingsGroupedByVulnerability(filters, showInactive);
+        return getFindingsSearchQueryManager().getAllFindingsGroupedByVulnerability(filters, showInactive);
     }
 
     public List<VulnerabilityMetrics> getVulnerabilityMetrics() {
